+++ src/LanguageServer.Engine/Documents/SubProjectDocument.cs
@@ -113,8 +113,6 @@ namespace MSBuildProjectTools.LanguageServer.Documents
                 Log.Error(loadError, "Error loading MSBuild project '{ProjectFileName}'.", ProjectFile.FullName);
             }
 
-            TryUnloadMSBuildProject();
-
             return false;
         }